Move layout to src folder
